requirements.txt
asgiref==3.2.10
bson==0.5.8
celery==5.0.2
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
coverage==5.3
cryptography==3.1.1
dataclasses==0.6
defusedxml==0.7.0rc1
Django==3.1.2
django-debug-toolbar==3.1.1
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-redis==4.12.1
djangorestframework==3.12.2
dnspython==2.0.0
gunicorn==20.0.4
idna==2.10
oauthlib==3.1.0
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.7.1
pymongo==3.11.2
python-dateutil==2.8.1
python3-openid==3.2.0
pytils==0.3
pytz==2020.1
redis==3.4.1
requests==2.24.0
requests-oauthlib==1.3.0
selenium==3.141.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.2.4
urllib3==1.25.10
whitenoise==5.2.0
